Remove old VC gen
