+++ frontends/common/src/it/scala/stainless/verification/VerificationSuite.scala
@@ -3,22 +3,17 @@
 package stainless
 package verification
 
+import scala.concurrent.duration._
+
 import org.scalatest._
 
 trait VerificationSuite extends VerificationComponentTestSuite {
 
-  override def configurations = super.configurations.map { seq =>
-    Seq(optTypeChecker(false)) ++ seq
-  }
-
-  override protected def optionsString(options: inox.Options): String = {
-    super.optionsString(options) +
-    (if (options.findOptionOrDefault(evaluators.optCodeGen)) " codegen" else "")
-  }
-
   override def filter(ctx: inox.Context, name: String): FilterStatus = name match {
-    case "verification/valid/Streams" => Ignore // only for TypeCheckerSuite
-    case "verification/invalid/ForallAssoc" => Ignore // Hangs
+    case "verification/invalid/ForallAssoc"           => Ignore  // Hangs
+
+    // unknown/timeout VC but counter-example not found
+    case "verification/invalid/BadConcRope"           => Ignore
 
     // Lemmas used in one equation can leak in other equations due to https://github.com/epfl-lara/inox/issues/139
     case "verification/invalid/Equations1" => Ignore
@@ -28,13 +23,25 @@ trait VerificationSuite extends VerificationComponentTestSuite {
     case _ => super.filter(ctx, name)
   }
 
-  testPosAll("verification/valid")
+  // Scala 2 BitVectors tests leverages the fact that `==` can be used to compare two unrelated types.
+  // For instance, if we have x: Int42, then x == 42 is legal.
+  // In Scala 3, however, this expression is ill-formed because Int42 and Int (the widened type of 42) are unrelated.
+  // Therefore, all BitVectors tests for Scala 3 must perform a conversion for literals
+  // (e.g. the above expression is rewritten to x == (42: Int42))
+  def bitVectorsTestDiscarding(file: String): Boolean = {
+    val scala2BitVectors = Set("MicroTests/scalac/BitVectors1.scala", "MicroTests/scalac/BitVectors2.scala", "MicroTests/scalac/BitVectors3.scala")
+    val scala3BitVectors = Set("MicroTests/dotty/BitVectors1.scala", "MicroTests/dotty/BitVectors2.scala", "MicroTests/dotty/BitVectors3.scala")
+
+    (isScala3 || !scala3BitVectors.exists(t => file.endsWith(t))) &&
+      (isScala2 || !scala2BitVectors.exists(t => file.endsWith(t)))
+  }
+
+  testPosAll("verification/valid", recursive = true, bitVectorsTestDiscarding)
 
   testNegAll("verification/invalid")
 
   // Tests that should be rejected, but aren't
   testPosAll("verification/false-valid")
-
 }
 
 class SMTZ3VerificationSuite extends VerificationSuite {
@@ -46,13 +53,6 @@ class SMTZ3VerificationSuite extends VerificationSuite {
   }
 
   override def filter(ctx: inox.Context, name: String): FilterStatus = name match {
-    // Flaky on smt-z3 for some reason
-    case "verification/valid/MergeSort2" => Ignore
-    case "verification/valid/IntSetInv" => Ignore
-
-    // Too slow on smt-z3, even for nightly build
-    case "verification/valid/BitsTricksSlow" => Skip
-
     case _ => super.filter(ctx, name)
   }
 }
@@ -81,41 +81,28 @@ class SMTCVC4VerificationSuite extends VerificationSuite {
   override def configurations = super.configurations.map {
     seq => Seq(
       inox.optSelectedSolvers(Set("smt-cvc4")),
-      inox.solvers.optCheckModels(true),
-      evaluators.optCodeGen(true)
+      inox.solvers.optCheckModels(true)
     ) ++ seq
   }
 
   override def filter(ctx: inox.Context, name: String): FilterStatus = name match {
-    // Requires non-linear resonning, unsupported by CVC4
-    case "verification/valid/Overrides" => Ignore
-    case "verification/valid/TestPartialFunction" => Ignore
-    case "verification/valid/TestPartialFunction3" => Ignore
-    case "verification/valid/BigIntRing" => Ignore
-    case "verification/valid/BigIntMonoidLaws" => Ignore
-    case "verification/valid/InnerClasses4" => Ignore
-
-    // assertion failed in `compileLambda`
-    case "verification/valid/GodelNumbering" => Ignore
-
-    // This test is flaky on CVC4
-    case "verification/valid/CovariantList" => Ignore
-
-    // Requires map with non-default values, unsupported by CVC4
     case "verification/valid/ArraySlice" => Ignore
-    case "verification/valid/Iterables" => Ignore
-
-    // These tests are too slow on CVC4 and make the regression unstable
+    case "verification/valid/BigIntMonoidLaws" => Ignore
+    case "verification/valid/BigIntRing" => Ignore
     case "verification/valid/ConcRope" => Ignore
-    case "verification/invalid/BadConcRope" => Ignore
-
-    // These tests make CVC4 crash
+    case "verification/valid/CovariantList" => Ignore
+    case "verification/valid/Huffman" => Ignore
+    case "verification/valid/InnerClasses4" => Ignore
+    case "verification/valid/Iterables" => Ignore
+    case "verification/valid/List" => Ignore
+    case "verification/valid/MoreExtendedEuclidGCD" => Ignore
+    case "verification/valid/MoreExtendedEuclidReachability" => Ignore
+    case "verification/valid/Overrides" => Ignore
     case "verification/valid/PartialCompiler" => Ignore
     case "verification/valid/PartialKVTrace" => Ignore
-
-    // Codegen assertion error, unsupported by CVC4
-    // => Ignored until #681 fixed
-    case "verification/invalid/BodyEnsuring" => Ignore
+    case "verification/valid/ReachabilityChecker" => Ignore
+    case "verification/valid/TestPartialFunction" => Ignore
+    case "verification/valid/TestPartialFunction3" => Ignore
 
     case _ => super.filter(ctx, name)
   }
